Sorted components in specific folders, added folder pages and changed App.tsx to only show the Header and the page
