+++ Frontend/profiles-of-endava/.gitignore
@@ -23,3 +23,5 @@ yarn-debug.log*
 yarn-error.log*
 
 .eslintcache
+
+debug.log
