+++ Frontend/profiles-of-endava/src/components/HeaderComponent/Header.scss
@@ -1,0 +1,14 @@
+*{
+    padding:0;
+    margin:0;
+}
+  
+
+$color-dark-grey:#666;
+$color-white:#FFF;
+
+
+.header{
+    background-color: $color-dark-grey;
+    color: $color-white;
+  }
