+++ Frontend/profiles-of-endava/src/components/PhotoAndInfoComponent/PhotoAndInfo.scss
@@ -1,0 +1,20 @@
+.photoAndInfo{
+    grid-row:1/2;
+    grid-column:1/4;
+    margin-bottom: 10px;
+    img{
+      float: left;
+      margin-right: 1em;
+    }
+    h2{
+      font-size: 20px;
+      margin-bottom: 10px;
+    }
+    h3{
+      font-size: 14px;
+      margin-bottom: 10px;
+    }
+    h4{
+      font-size: 10px;
+    }
+  }
