+++ Frontend/profiles-of-endava/src/components/PhotoAndInfoComponent/PhotoAndInfo.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import "../PhotoAndInfoComponent/PhotoAndInfo.scss"
 
 interface PhotoAndInfoProps {
     image?:string,
