+++ Frontend/profiles-of-endava/src/components/SkillsComponent/Skills.scss
@@ -1,0 +1,28 @@
+$color-blue:#08619D;
+$color-dark-grey:#666;
+
+.skills{
+    grid-row:2/3;
+    grid-column:1/4;
+    h3{
+        border-bottom: 1px solid $color-blue;
+      }
+    ul{
+    list-style-type: none;
+    margin-top: 10px;
+    float:left;
+    padding: 0 55px 0 10px;
+    font-size:14px;
+    }
+    ul:nth-child(2){
+    padding-left:0;
+    }
+    ul:nth-child(3){
+    border-right:1px solid $color-dark-grey;
+    border-left:1px solid $color-dark-grey;
+    }
+    li{
+    margin:5px 0;
+    font-size:1em;
+    }
+}
