+++ Frontend/profiles-of-endava/src/styles/App.scss
@@ -13,78 +13,3 @@ $color-dark-grey:#666;
   background-color: $color-light-grey;
 }
 
-.header{
-  background-color: $color-dark-grey;
-  color: $color-white;
-}
-
-.card{
-  width:379px;
-  height:381px;
-  border:1px solid grey;
-  padding:22px;
-  box-sizing:border-box;
-  display:grid;
-  grid-template:repeat(3, 1fr)/repeat(3,1fr);
-  background-color:$color-white;
-}
-
-.photoAndInfo{
-  grid-row:1/2;
-  grid-column:1/4;
-  margin-bottom: 10px;
-  img{
-    float: left;
-    margin-right: 1em;
-  }
-  h2{
-    font-size: 20px;
-    margin-bottom: 10px;
-  }
-  h3{
-    font-size: 14px;
-    margin-bottom: 10px;
-  }
-  h4{
-    font-size: 10px;
-  }
-}
-
-.skills{
-  grid-row:2/3;
-  grid-column:1/4;
-}
-
-.projects{
-  grid-row:3/4;
-  grid-column:1/4;
-  margin-top: 5px;
-}
-
-h3{
-  color:$color-blue;
-}
-
-.skills, .projects{
-  h3{
-    border-bottom: 1px solid $color-blue;
-  }
-  ul{
-    list-style-type: none;
-    margin-top: 10px;
-    float:left;
-    padding: 0 55px 0 10px;
-    font-size:14px;
-  }
-  ul:nth-child(2){
-    padding-left:0;
-  }
-  ul:nth-child(3){
-    border-right:1px solid $color-dark-grey;
-    border-left:1px solid $color-dark-grey;
-  }
-  li{
-    margin:5px 0;
-    font-size:1em;
-  }
-}
